requirements/dev.txt
-r common.txt

Flask-Testing==0.7.1
pytest==3.6.3
pytest-mock==1.10.0
python-dotenv==0.8.2

coverage==4.5.1
flake8==3.5.0
mypy==0.620
httmock==1.2.6
black==18.6b4
